nixbld-etc-nixos/rfq/src/requirements.txt
blinker
click
Flask
Flask-Mail
itsdangerous
Jinja2
MarkupSafe
python-dotenv
six
Werkzeug
